requirements.txt
Django
python-dotenv
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
supabase
psycopg2
psycopg2-binary
flake8
google-cloud-firestore